+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.19.2] - 2026-08-24
+
 ### Changed
 
 - **A table view declaring `order_by` now fails as the class is defined**, rather